move FullWidthListItem and FullWidthListContainer to the components folder
